backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pyserial==3.5
websockets==13.1
psutil==6.1.1
